Move tests in their respective components
