+++ package.json
@@ -47,7 +47,7 @@
     "preset": "ts-jest",
     "testEnvironment": "node",
     "testMatch": [
-      "**/tests/**/*.test.ts"
+      "**/src/**/*.test.ts"
     ]
   }
 }
